public/machine-learning/requirements.txt
blinker==1.6.2
click==8.1.7
Flask==2.3.3
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
MarkupSafe==2.1.3
numpy==1.24.4
pandas==2.0.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
scikit-learn==1.3.0
scipy==1.10.1
six==1.16.0
threadpoolctl==3.2.0
tzdata==2023.3
Werkzeug==2.3.7
zipp==3.17.0
